backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
anthropic==0.34.2
numpy
scikit-learn
python-dotenv
